+++ R/snk.evaluate.core.R
@@ -32,16 +32,17 @@
 #'   core collection and present in the \code{names} column.
 #'
 #' @return A data frame with the following components. \item{Trait}{The
-#'   quantitative trait.} \item{EC_Min}{The minimum value of the trait in EC.}
-#'   \item{EC_Max}{The maximum value of the trait in EC.} \item{EC_Mean}{The
-#'   mean value of the trait in EC.} \item{EC_SE}{The standard error of the
-#'   trait in EC.} \item{CS_Min}{The minimum value of the trait in CS.}
-#'   \item{CS_Max}{The maximum value of the trait in CS.} \item{CS_Mean}{The
-#'   mean value of the trait in CS.} \item{CS_SE}{The standard error of the
-#'   trait in CS.} \item{SNK_pvalue}{The p value of the Student-Newman-Keuls
-#'   test for equality of means of EC and CS.} \item{SNK_significance}{The
-#'   significance of the Student-Newman-Keuls test for equality of means of EC
-#'   and CS.}
+#'   quantitative trait.} \item{Count}{The accession count (excluding missing
+#'   data).} \item{Df}{The degrees of freedom for the test.} \item{EC_Min}{The
+#'   minimum value of the trait in EC.} \item{EC_Max}{The maximum value of the
+#'   trait in EC.} \item{EC_Mean}{The mean value of the trait in EC.}
+#'   \item{EC_SE}{The standard error of the trait in EC.} \item{CS_Min}{The
+#'   minimum value of the trait in CS.} \item{CS_Max}{The maximum value of the
+#'   trait in CS.} \item{CS_Mean}{The mean value of the trait in CS.}
+#'   \item{CS_SE}{The standard error of the trait in CS.} \item{SNK_pvalue}{The
+#'   p value of the Student-Newman-Keuls test for equality of means of EC and
+#'   CS.} \item{SNK_significance}{The significance of the Student-Newman-Keuls
+#'   test for equality of means of EC and CS.}
 #'
 #' @seealso \code{\link[agricolae]{SNK.test}}
 #'
@@ -107,9 +108,10 @@ snk.evaluate.core <- function(data, names, quantitative, selected) {
   for (i in seq_along(quantitative)) {
     frmla <- stats::formula(paste("`", quantitative[i], "` ~ `[Type]`",
                                        sep = ""))
-    model <- stats::aov(frmla, data = dataf)
+    model <- stats::aov(frmla, data = dataf[!is.na(dataf[, quantitative[i]]), ])
     snkout <- agricolae::SNK.test(model, "[Type]", group = FALSE,
                                   console = FALSE)
+    snkdf <- snkout$statistics$Df
     snkpvalue <- snkout$comparison$pvalue
 
     # out <- mutoss::snk(frmla, data = dataf,
@@ -117,17 +119,39 @@ snk.evaluate.core <- function(data, names, quantitative, selected) {
     # out <- t.test(dataf[dataf$`[Type]` == "EC", quantitative[i]],
     #               dataf[dataf$`[Type]` == "CS", quantitative[i]])
 
-    outdf[[quantitative[i]]] <- data.frame(`Trait` = quantitative[i],
-                                           `EC_Min` = min(dataf[dataf$`[Type]` == "EC", quantitative[i]]),
-                                           `EC_Max` = max(dataf[dataf$`[Type]` == "EC", quantitative[i]]),
-                                           `EC_Mean` = mean(dataf[dataf$`[Type]` == "EC", quantitative[i]]),
-                                           `EC_SE` = stats::sd(dataf[dataf$`[Type]` == "EC", quantitative[i]]) / sqrt(length(dataf[dataf$`[Type]` == "EC", quantitative[i]])),
-                                           `CS_Min` = min(dataf[dataf$`[Type]` == "CS", quantitative[i]]),
-                                           `CS_Max` = max(dataf[dataf$`[Type]` == "CS", quantitative[i]]),
-                                           `CS_Mean` = mean(dataf[dataf$`[Type]` == "CS", quantitative[i]]),
-                                           `CS_SE` = stats::sd(dataf[dataf$`[Type]` == "CS", quantitative[i]]) / sqrt(length(dataf[dataf$`[Type]` == "CS", quantitative[i]])),
-                                           `SNK_pvalue` = snkpvalue,
-                                           stringsAsFactors = FALSE)
+    outdf[[quantitative[i]]] <-
+      data.frame(`Trait` = quantitative[i],
+                 `Count` = sum(!is.na(dataf[dataf$`[Type]` == "EC",
+                                            quantitative[i]])),
+                 `Df` = snkdf,
+                 `EC_Min` = min(dataf[dataf$`[Type]` == "EC", quantitative[i]],
+                                na.rm = TRUE),
+                 `EC_Max` = max(dataf[dataf$`[Type]` == "EC", quantitative[i]],
+                                na.rm = TRUE),
+                 `EC_Mean` = mean(dataf[dataf$`[Type]` == "EC",
+                                        quantitative[i]],
+                                  na.rm = TRUE),
+                 `EC_SE` = stats::sd(dataf[dataf$`[Type]` == "EC",
+                                           quantitative[i]],
+                                     na.rm = TRUE) /
+                   sqrt(length(dataf[dataf$`[Type]` == "EC" &
+                                       !is.na(dataf[, quantitative[i]]),
+                                     quantitative[i]])),
+                 `CS_Min` = min(dataf[dataf$`[Type]` == "CS", quantitative[i]],
+                                na.rm = TRUE),
+                 `CS_Max` = max(dataf[dataf$`[Type]` == "CS", quantitative[i]],
+                                na.rm = TRUE),
+                 `CS_Mean` = mean(dataf[dataf$`[Type]` == "CS",
+                                        quantitative[i]],
+                                  na.rm = TRUE),
+                 `CS_SE` = stats::sd(dataf[dataf$`[Type]` == "CS",
+                                           quantitative[i]],
+                                     na.rm = TRUE) /
+                   sqrt(length(dataf[dataf$`[Type]` == "CS" &
+                                       !is.na(dataf[, quantitative[i]]),
+                                     quantitative[i]])),
+                 `SNK_pvalue` = snkpvalue,
+                 stringsAsFactors = FALSE)
 
     rm(snkout, snkpvalue, frmla, model)
   }